requirements.txt
# For pyfreeradius module
pydantic

# All supported DB-API 2.0 (PEP 249) drivers
mysql-connector-python
pymssql
psycopg2
#sqlite3  # already included in Python standard library
oracledb

# For FastAPI app
fastapi
uvicorn
pydantic-settings